backend/requirements.txt
# AI-PatternWeb Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==24.1.0
redis==5.0.8
pydantic==2.9.0
pydantic-settings==2.5.0
google-generativeai==0.8.0
Pillow==10.4.0
shapely==2.0.5
ezdxf==0.19.0
reportlab==4.2.0
python-dotenv==1.0.1
httpx==0.27.2
celery==5.4.0
email-validator==2.2.0
